Apply changes in admin panel related with refactor code separated in components field
