+++ client/cards/src/components/admin/adminUtils/orders/ShowOrders.module.css
@@ -1,0 +1,64 @@
+.tableRow {
+    display: flex;
+    flex-flow: row nowrap;
+    width: 100%;
+    border-bottom: 1px solid #dadada;
+}
+
+.tableRow:last-child,
+.rowSubContainer .row-item:last-child {
+    border-bottom: 0;
+}
+
+.tableRow .rowItem {
+    display: flex;
+    flex: 1;
+    font-size: 12px;
+    padding: 8px 0;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    transition: all 0.15s ease-in-out;
+}
+
+.tableRow .rowItem:hover {
+    cursor: pointer;
+    background-color: #F0F0F0;
+}
+
+.tableRow .rowItemOrder {
+    display: flex;
+    flex: 1;
+    font-size: 12px;
+    padding: 8px 0;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    transition: all 0.15s ease-in-out;
+}
+
+.buttonSeeItemOrder {
+    all: unset;
+    width: 80%;
+    height: 80%;
+    border-radius: 5px;
+    font-weight: 600;
+    cursor: pointer;
+    background-color: #b0dcf44a;
+    color: rgb(4, 78, 20);
+}
+
+.tableRow .buttonSeeItemOrder:hover {
+    background-color: rgba(96, 175, 235, 0.317);
+}
+
+.tableRow .rowSubContainer {
+    display: flex;
+    flex-flow: column nowrap;
+    flex: 1;
+}
+
+.tableRow .rowSubContainer .row-item {
+    padding: 8px 0;
+    border-bottom: 1px solid #dadada;
+}
