+++ client/cards/src/components/admin/adminUtils/orders/rowItems/RowItems.module.css
@@ -1,0 +1,23 @@
+.rowItem {
+    display: flex;
+    flex: 1;
+    font-size: 12px;
+    padding: 8px 0;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    transition: all 0.15s ease-in-out;
+}
+
+.rowItem:hover {
+    cursor: pointer;
+    background-color: #F0F0F0;
+}
+
+.tooltipMessage {
+    display: flex;
+}
+
+.messageIcon {
+    width: 30px;
+}
