+++ client/cards/src/components/admin/allOrders/AllOrders.module.css
@@ -6,8 +6,6 @@
     background-color: #fafafa;
     border-top-right-radius: 10px;
     border-top-left-radius: 10px;
-
-    /* border: 1px solid black; */
 }
 
 .tableContainer {
@@ -16,86 +14,12 @@
     background-color: white;
     width: 100%;
     margin: 0 auto;
-    /* border-radius: 4px; */
     border: 1px solid #DADADA;
     box-shadow: 0px 1px 4px rgba(0, 0, 0, .08);
-}
-
-.tableRow {
-    display: flex;
-    flex-flow: row nowrap;
-    width: 100%;
-    border-bottom: 1px solid #dadada;
 }
 
 .heading {
     background-color: #ececec;
     color: #3e3e3e;
     font-weight: bold;
-}
-
-.rowItem {
-    display: flex;
-    flex: 1;
-    font-size: 12px;
-    padding: 8px 0;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    transition: all 0.15s ease-in-out;
-}
-
-.rowItem:hover {
-    cursor: pointer;
-    background-color: #F0F0F0;
-}
-
-.tooltipMessage {
-    display: flex;
-}
-
-.rowSubContainer {
-    display: flex;
-    flex-flow: column nowrap;
-    flex: 1;
-}
-
-.rowSubContainer .row-item {
-    padding: 8px 0;
-    border-bottom: 1px solid #dadada;
-}
-
-.tableRow:last-child,
-.rowSubContainer .row-item:last-child {
-    border-bottom: 0;
-}
-
-.messageIcon {
-    width: 30px;
-}
-
-.rowItemOrder {
-    display: flex;
-    flex: 1;
-    font-size: 12px;
-    padding: 8px 0;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    transition: all 0.15s ease-in-out;
-}
-
-.buttonSeeItemOrder {
-    all: unset;
-    width: 80%;
-    height: 80%;
-    border-radius: 5px;
-    font-weight: 600;
-    cursor: pointer;
-    background-color: #b0dcf44a;
-    color: rgb(4, 78, 20);
-}
-
-.buttonSeeItemOrder:hover {
-    background-color: rgba(96, 175, 235, 0.317);
 }
